+++ task_test.go
@@ -10,6 +10,7 @@ import (
 	"runtime"
 	"strings"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 
@@ -476,6 +477,55 @@ func TestStatusChecksum(t *testing.T) {
 	assert.Equal(t, `task: Task "build" is up to date`+"\n", buff.String())
 }
 
+func TestAlias(t *testing.T) {
+	const dir = "testdata/alias"
+
+	data, err := os.ReadFile(filepathext.SmartJoin(dir, "alias.txt"))
+	assert.NoError(t, err)
+
+	var buff bytes.Buffer
+	e := task.Executor{
+		Dir:    dir,
+		Stdout: &buff,
+		Stderr: &buff,
+	}
+	assert.NoError(t, e.Setup())
+	assert.NoError(t, e.Run(context.Background(), taskfile.Call{Task: "f"}))
+	assert.Equal(t, string(data), buff.String())
+}
+
+func TestDuplicateAlias(t *testing.T) {
+	const dir = "testdata/alias"
+
+	var buff bytes.Buffer
+	e := task.Executor{
+		Dir:    dir,
+		Stdout: &buff,
+		Stderr: &buff,
+	}
+	assert.NoError(t, e.Setup())
+	assert.Error(t, e.Run(context.Background(), taskfile.Call{Task: "x"}))
+	assert.Equal(t, "", buff.String())
+}
+
+func TestAliasSummary(t *testing.T) {
+	const dir = "testdata/alias"
+
+	data, err := os.ReadFile(filepathext.SmartJoin(dir, "alias-summary.txt"))
+	assert.NoError(t, err)
+
+	var buff bytes.Buffer
+	e := task.Executor{
+		Dir:     dir,
+		Summary: true,
+		Stdout:  &buff,
+		Stderr:  &buff,
+	}
+	assert.NoError(t, e.Setup())
+	assert.NoError(t, e.Run(context.Background(), taskfile.Call{Task: "f"}))
+	assert.Equal(t, string(data), buff.String())
+}
+
 func TestLabelUpToDate(t *testing.T) {
 	const dir = "testdata/label_uptodate"
 
@@ -546,7 +596,7 @@ func TestLabelInSummary(t *testing.T) {
 	assert.Contains(t, buff.String(), "foobar")
 }
 
-func TestLabelInList(t *testing.T) {
+func TestNoLabelInList(t *testing.T) {
 	const dir = "testdata/label_list"
 
 	var buff bytes.Buffer
@@ -556,8 +606,8 @@ func TestLabelInList(t *testing.T) {
 		Stderr: &buff,
 	}
 	assert.NoError(t, e.Setup())
-	e.ListTasksWithDesc()
-	assert.Contains(t, buff.String(), "foobar")
+	e.ListTasks(task.FilterOutInternal(), task.FilterOutNoDesc())
+	assert.Contains(t, buff.String(), "foo")
 }
 
 // task -al case 1: listAll list all tasks
@@ -574,7 +624,7 @@ func TestListAllShowsNoDesc(t *testing.T) {
 	assert.NoError(t, e.Setup())
 
 	var title string
-	e.ListAllTasks()
+	e.ListTasks(task.FilterOutInternal())
 	for _, title = range []string{
 		"foo",
 		"voo",
@@ -596,7 +646,7 @@ func TestListCanListDescOnly(t *testing.T) {
 	}
 
 	assert.NoError(t, e.Setup())
-	e.ListTasksWithDesc()
+	e.ListTasks(task.FilterOutInternal(), task.FilterOutNoDesc())
 
 	var title string
 	assert.Contains(t, buff.String(), "foo")
@@ -838,6 +888,22 @@ func TestIncorrectVersionIncludes(t *testing.T) {
 	assert.EqualError(t, e.Setup(), expectedError)
 }
 
+func TestIncludesIncorrect(t *testing.T) {
+	const dir = "testdata/includes_incorrect"
+
+	var buff bytes.Buffer
+	e := task.Executor{
+		Dir:    dir,
+		Stdout: &buff,
+		Stderr: &buff,
+		Silent: true,
+	}
+
+	err := e.Setup()
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "task: Failed to parse testdata/includes_incorrect/incomplete.yml:")
+}
+
 func TestIncludesEmptyMain(t *testing.T) {
 	tt := fileContentTest{
 		Dir:       "testdata/includes_empty",
@@ -968,12 +1034,7 @@ func TestIncludesInternal(t *testing.T) {
 	}{
 		{"included internal task via task", "task-1", false, "Hello, World!\n"},
 		{"included internal task via dep", "task-2", false, "Hello, World!\n"},
-		{
-			"included internal direct",
-			"included:task-3",
-			true,
-			"task: No tasks with description available. Try --list-all to list all tasks\n",
-		},
+		{"included internal direct", "included:task-3", true, ""},
 	}
 
 	for _, test := range tests {
@@ -989,7 +1050,7 @@ func TestIncludesInternal(t *testing.T) {
 
 			err := e.Run(context.Background(), taskfile.Call{Task: test.task})
 			if test.expectedErr {
-				assert.Error(t, err, test.expectedErr)
+				assert.Error(t, err)
 			} else {
 				assert.NoError(t, err)
 			}
@@ -1008,12 +1069,7 @@ func TestInternalTask(t *testing.T) {
 	}{
 		{"internal task via task", "task-1", false, "Hello, World!\n"},
 		{"internal task via dep", "task-2", false, "Hello, World!\n"},
-		{
-			"internal direct",
-			"task-3",
-			true,
-			"task: No tasks with description available. Try --list-all to list all tasks\n",
-		},
+		{"internal direct", "task-3", true, ""},
 	}
 
 	for _, test := range tests {
@@ -1029,13 +1085,37 @@ func TestInternalTask(t *testing.T) {
 
 			err := e.Run(context.Background(), taskfile.Call{Task: test.task})
 			if test.expectedErr {
-				assert.Error(t, err, test.expectedErr)
+				assert.Error(t, err)
 			} else {
 				assert.NoError(t, err)
 			}
 			assert.Equal(t, test.expectedOutput, buff.String())
 		})
 	}
+}
+
+func TestIncludesShadowedDefault(t *testing.T) {
+	tt := fileContentTest{
+		Dir:       "testdata/includes_shadowed_default",
+		Target:    "included",
+		TrimSpace: true,
+		Files: map[string]string{
+			"file.txt": "shadowed",
+		},
+	}
+	tt.Run(t)
+}
+
+func TestIncludesUnshadowedDefault(t *testing.T) {
+	tt := fileContentTest{
+		Dir:       "testdata/includes_unshadowed_default",
+		Target:    "included",
+		TrimSpace: true,
+		Files: map[string]string{
+			"file.txt": "included",
+		},
+	}
+	tt.Run(t)
 }
 
 func TestSupportedFileNames(t *testing.T) {
@@ -1478,5 +1558,66 @@ func TestEvaluateSymlinksInPaths(t *testing.T) {
 	assert.NoError(t, err)
 	buff.Reset()
 	err = os.RemoveAll(dir + "/.task")
+	assert.NoError(t, err)
+}
+
+func TestFileWatcherInterval(t *testing.T) {
+	const dir = "testdata/watcher_interval"
+	expectedOutput := strings.TrimSpace(`
+task: Started watching for tasks: default
+task: [default] echo "Hello, World!"
+Hello, World!
+task: [default] echo "Hello, World!"
+Hello, World!
+	`)
+
+	var buff bytes.Buffer
+	e := &task.Executor{
+		Dir:    dir,
+		Stdout: &buff,
+		Stderr: &buff,
+		Watch:  true,
+	}
+
+	assert.NoError(t, e.Setup())
+	buff.Reset()
+
+	err := os.MkdirAll(filepathext.SmartJoin(dir, "src"), 0755)
+	assert.NoError(t, err)
+
+	err = os.WriteFile(filepathext.SmartJoin(dir, "src/a"), []byte("test"), 0644)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	ctx := context.Background()
+	ctx, cancel := context.WithCancel(ctx)
+
+	go func(ctx context.Context) {
+		for {
+			select {
+			case <-ctx.Done():
+				return
+			default:
+				err := e.Run(ctx, taskfile.Call{Task: "default"})
+				if err != nil {
+					return
+				}
+			}
+		}
+	}(ctx)
+
+	time.Sleep(10 * time.Millisecond)
+	err = os.WriteFile(filepathext.SmartJoin(dir, "src/a"), []byte("test updated"), 0644)
+	if err != nil {
+		t.Fatal(err)
+	}
+	time.Sleep(700 * time.Millisecond)
+	cancel()
+	assert.Equal(t, expectedOutput, strings.TrimSpace(buff.String()))
+	buff.Reset()
+	err = os.RemoveAll(filepathext.SmartJoin(dir, ".task"))
+	assert.NoError(t, err)
+	err = os.RemoveAll(filepathext.SmartJoin(dir, "src"))
 	assert.NoError(t, err)
 }